GROW-4 fix : stock_daily/weekly/monthly, id 필드 추가 및 데이터 수집 테스트

+++ backend/app/module/auth/model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, DateTime, Enum, Integer, String
+from sqlalchemy import BigInteger, Column, DateTime, Enum, String
 from sqlalchemy.orm import relationship
 
 from app.common.mixin.timestamp import TimestampMixin
@@ -9,7 +9,7 @@ from database.config import MySQLBase
 class User(TimestampMixin, MySQLBase):
     __tablename__ = "user"
 
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
     social_id = Column(String(255), index=True, nullable=False)
     provider = Column(String(50), nullable=False)
     role = Column(Enum(UserRoleEnum), default=UserRoleEnum.USER)
